Update field type for OrderComponent cost from float(2,2) to decimal(8,2)

+++ database/migrations/2021_12_13_175840_create_order_merchandises_table.php
@@ -17,7 +17,7 @@ class CreateOrderMerchandisesTable extends Migration
             $table->id();
             $table->foreignId('order_customer_id')->constrained()->onDelete('cascade');
             $table->foreignId('merchandise_id')->constrained()->onDelete('cascade');
-            $table->float('cost', 2);
+            $table->decimal('cost');
             $table->timestamps();
             $table->softDeletes();
         });
